requirements.txt
basemap==1.0.7
beautifulsoup4==4.5.0
cycler==0.10.0
get==0.0.11
html5lib==0.999999999
lxml==3.6.1
matplotlib==1.5.1
numpy==1.11.1
pandas==0.18.1
Pillow==3.3.0
post==0.0.8
public==0.0.34
pyparsing==2.1.5
python-dateutil==2.5.3
pytz==2016.6.1
query-string==0.0.8
request==0.0.8
requests==2.10.0
scikit-learn==0.17.1
six==1.10.0
virtualenv==15.0.2
webencodings==0.5
